Move `QuestionDescription` component
