+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/QuestionDescription/index.ts
@@ -1,0 +1,1 @@
+export { QuestionDescription } from "./QuestionDescription";
